Support ^[t] tags for array type hinting
This patch changes the behavior of the reader so that ^[..] is a valid
metadata form, and defines it to be a hinting tag same as ^int or
^float. Such hints must be of one element in length, where that element
is a symbol or a string denoting a type.

This patch modifies the way that the compiler interprets type tags, so
that vectors of one element will be interpreted recursively to mean an
array of that type. So rather than having to hint ^"[Ljava.lang.Object;"
or the pun therefor, ^objects, a user may now simply hint
^[Object]. This notation also nests, so ^[[Object]] or ^[[int]] behaves
as one would expect.